wink_light: Implement getstate.
Returns the state of a Wink light, such as ON or OFF. Also
adds its intensity. Exempli gratia:

  $ akasha winkbulb getstate
  ON 0.50

This indicates a light bulb turned on at 50% brightness.

Note: the OFF state also indicates a brightness, as simply
switching it on would set it to such level of brightness.
E.g.:

  $ akasha winkbulb poweroff
  $ akasha winkbulb getstate
  OFF 0.50

If a light bulb is unreachable, or if there's any other
kind of unexpected problem, the returned state will be an
error and a brief explanation:
  $ akasha winkbulb getstate
  ERR Not connected

Signed-off-by: Daniel Díaz <yosoy@danieldiaz.org>