cuby: Report on correct Cuby device id.
The API returns the information of _ALL_ the Cuby devices
when requesting data. When more than one Cuby is assigned
to a user, only the first one would be considered by Akasha.

This fixes that by traversing the collection of Cuby devices
one by one and finding the right information.

Fixes #7.

Signed-off-by: Daniel Díaz <yosoy@danieldiaz.org>